+++ src/main/java/kitchenpos/menu/application/MenuService.java
@@ -10,9 +10,9 @@ import kitchenpos.menu.application.dto.MenuResponse;
 import kitchenpos.menu.domain.repository.MenuDao;
 import kitchenpos.menu.domain.repository.MenuGroupDao;
 import kitchenpos.menu.domain.repository.MenuProductDao;
-import kitchenpos.product.domain.repository.ProductDao;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuProduct;
+import kitchenpos.product.domain.repository.ProductDao;
 import kitchenpos.product.domain.Product;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
