+++ src/main/java/kitchenpos/table/domain/repository/TableGroupDao.java
@@ -1,8 +1,8 @@
-package kitchenpos.tableGroup.repository;
+package kitchenpos.table.domain.repository;
 
 import java.util.List;
 import java.util.Optional;
-import kitchenpos.tableGroup.domain.TableGroup;
+import kitchenpos.table.domain.TableGroup;
 
 public interface TableGroupDao {
     TableGroup save(TableGroup entity);
