+++ src/test/java/kitchenpos/application/OrderServiceTest.java
@@ -15,14 +15,14 @@ import kitchenpos.menu.domain.repository.MenuDao;
 import kitchenpos.menu.domain.repository.MenuGroupDao;
 import kitchenpos.order.domain.repository.OrderDao;
 import kitchenpos.order.domain.repository.OrderLineItemDao;
-import kitchenpos.order.domain.repository.OrderTableDao;
+import kitchenpos.table.domain.repository.OrderTableDao;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuGroup;
 import kitchenpos.order.application.OrderService;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderLineItem;
 import kitchenpos.order.domain.OrderStatus;
-import kitchenpos.order.domain.OrderTable;
+import kitchenpos.table.domain.OrderTable;
 import kitchenpos.support.ServiceTest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
