+++ src/test/java/kitchenpos/application/TableGroupServiceTest.java
@@ -7,17 +7,17 @@ import java.time.LocalDateTime;
 import java.util.Collections;
 import java.util.List;
 import java.util.stream.Stream;
-import kitchenpos.order.application.dto.OrderTableRequest;
-import kitchenpos.tableGroup.application.TableGroupService;
-import kitchenpos.tableGroup.application.dto.TableGroupRequest;
-import kitchenpos.tableGroup.application.dto.TableGroupResponse;
+import kitchenpos.table.application.dto.OrderTableRequest;
+import kitchenpos.table.application.TableGroupService;
+import kitchenpos.table.application.dto.TableGroupRequest;
+import kitchenpos.table.application.dto.TableGroupResponse;
 import kitchenpos.order.domain.repository.OrderDao;
-import kitchenpos.order.domain.repository.OrderTableDao;
-import kitchenpos.tableGroup.repository.TableGroupDao;
+import kitchenpos.table.domain.repository.OrderTableDao;
+import kitchenpos.table.domain.repository.TableGroupDao;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderStatus;
-import kitchenpos.order.domain.OrderTable;
-import kitchenpos.tableGroup.domain.TableGroup;
+import kitchenpos.table.domain.OrderTable;
+import kitchenpos.table.domain.TableGroup;
 import kitchenpos.support.ServiceTest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
