+++ src/test/java/kitchenpos/application/TableServiceTest.java
@@ -4,17 +4,17 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.time.LocalDateTime;
-import kitchenpos.order.application.TableService;
-import kitchenpos.order.application.dto.OrderTableRequest;
-import kitchenpos.order.application.dto.OrderTableRequest.NumberOfGuest;
-import kitchenpos.order.application.dto.OrderTableResponse;
+import kitchenpos.table.application.TableService;
+import kitchenpos.table.application.dto.OrderTableRequest;
+import kitchenpos.table.application.dto.OrderTableRequest.NumberOfGuest;
+import kitchenpos.table.application.dto.OrderTableResponse;
 import kitchenpos.order.domain.repository.OrderDao;
-import kitchenpos.order.domain.repository.OrderTableDao;
-import kitchenpos.tableGroup.repository.TableGroupDao;
+import kitchenpos.table.domain.repository.OrderTableDao;
+import kitchenpos.table.domain.repository.TableGroupDao;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderStatus;
-import kitchenpos.order.domain.OrderTable;
-import kitchenpos.tableGroup.domain.TableGroup;
+import kitchenpos.table.domain.OrderTable;
+import kitchenpos.table.domain.TableGroup;
 import kitchenpos.support.ServiceTest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
